requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openapi
beautifulsoup4
streamlit
langchain-ollama
